Add \ to end of namespace in autoload
As the class has the same name as the root namespace, without
this \ suffix, the autoloader would try to load CustomBulkAction/.php
if one forgot to define a use statment, this would cause for a very
cryptic error message. Not any more though, now you get a "ClassNotFound"
error without "use". As one would expect